plug-ins: better support of big images: quality checkbox, loading notice, cleaner exit, frames locking.
The "preview quality" checkbox allows to preview the animation as at
most half the original image's dimension. Big animations (more than the
screen's dimension) with many frames will be checked by default.
This allows to significantly reduce loading time for huge images.

While loading, the playing progress bar will be used as a loading
progress bar, and care has been taken so that the UI is not frozen when
this happens. This allows a much better user experience.

Also the created frame image must be cleaned on exit, because even
though the process will finish, GEGL won't know about it (bug 690512).

Finally I lock the UI and prevent several actions on frames in the same
time, like rendering, or reinitialization.